3_Tool/PythonCan/requirement.txt
python-can==3.3.4
uptime==3.0.1